+++ mcp/langchain/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @metorial/langchain
 
+## 2.0.1
+
+### Patch Changes
+
+- Fixes for session creation with template and langchain sdk
+- Updated dependencies
+  - @metorial/mcp-session@2.0.1
+
 ## 2.0.0
 
 ### Major Changes