+++ mcp/mcp-session/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @metorial/mcp-session
 
+## 2.0.1
+
+### Patch Changes
+
+- Fixes for session creation with template and langchain sdk
+
 ## 2.0.0
 
 ### Major Changes